issue17: preview images now shown in fullheight and is adapted to window size
